move logic radio button to index
